app and main page organized
