+++ src/components/App/App.jsx
@@ -1,36 +1,29 @@
-import React, { useState, useEffect, useContext } from "react";
-import { useLocation } from "react-router-dom";
+import React, { useState, useEffect } from "react";
+import { useLocation, Route, Routes } from "react-router-dom";
 import "./App.css";
-import backgroundImg from "../../assets/background.svg";
-import notfoundImg from "../../assets/notfound.svg";
-import Header from "../../components/Header/Header";
-import Preloader from "../Preloader/Preloader";
-import About from "../About/About";
-import NewsGrid from "../NewsGrid/NewsGrid";
+
 import LoginModal from "../LoginModal/LoginModal";
 import RegisterModal from "../RegisterModal/RegisterModal";
 import SuccessRegisterModal from "../SucessRegisterModal/SuccessRegisterModal";
-import { fetchNews } from "../../utils/api";
-import Footer from "../Footer/Footer";
 import { authorize, checkToken } from "../../utils/Auth";
-import { CurrentUserContext } from "../Contexts/CurrentUserContexts";
-import { Route, Routes } from "react-router-dom";
+
+import Footer from "../Footer/Footer";
 import SaveArticles from "../SaveArticles/SaveArticles";
+import Main from "../Main/Main";
+import Navigation from "../Navigation/Navigation";
+
+import { CurrentUserContext } from "../Contexts/CurrentUserContexts";
 import { CurrentLocationContext } from "../Contexts/CurrentLocationContexts";
+
 const App = () => {
   const [activeModal, setActiveModal] = useState("");
   const [isLoggedIn, setIsLoggedIn] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-  const [searchResults, setSearchResults] = useState([]);
-  const [searchQuery, setSearchQuery] = useState("");
-  const [currentPage, setCurrentPage] = useState(1);
-  const [totalResults, setTotalResults] = useState(0);
-  const [error, setError] = useState(null);
   const [currentUser, setCurrentUser] = useState(null);
   const [currentLocation, setCurrentLocation] = useState("/");
 
   const location = useLocation();
   console.log("location", location);
+
   const handleLogin = ({ email, password }) => {
     authorize(email, password).then((data) => {
       checkToken(data.token).then((data) => {
@@ -40,50 +33,16 @@ const App = () => {
       });
     });
   };
+
   const handleLogout = () => {
     setIsLoggedIn(false);
     setCurrentUser(null);
   };
 
-  const handleSearch = async (query) => {
-    if (!query.trim()) return;
-
-    setSearchQuery(query);
-    setIsLoading(true);
-    setCurrentPage(1);
-    try {
-      const { articles, totalResults } = await fetchNews(query, 3, 1);
-      setSearchResults(articles);
-      setTotalResults(totalResults);
-      setError(null);
-    } catch (err) {
-      setError("Error fetching news. Please try again.");
-      setSearchResults([]);
-      setTotalResults(0);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleShowMore = async () => {
-    setIsLoading(true);
-    try {
-      const nextPage = currentPage + 1;
-      const { articles } = await fetchNews(searchQuery, 3, nextPage);
-      setSearchResults((prev) => [...prev, ...articles]);
-      setCurrentPage(nextPage);
-      setError(null);
-    } catch (err) {
-      setError("Error loading more articles. Please try again.");
-    } finally {
-      setIsLoading(false);
-    }
-  };
-  const moreArticles = searchResults.length < totalResults;
-
   const closeActiveModal = () => {
     setActiveModal("");
   };
+
   const handleSigninClick = () => {
     if (isLoggedIn) {
       handleLogout();
@@ -91,18 +50,19 @@ const App = () => {
       setActiveModal("signin");
     }
   };
+
   const handleSignupClick = () => {
     setActiveModal("signup");
   };
+
   const handleSuccessRegistration = () => {
     setActiveModal("success");
   };
+
   const handleCloseClick = () => {
     setActiveModal("");
   };
 
-  // Article page
-  useEffect(() => {});
   useEffect(() => {
     if (!activeModal) return;
 
@@ -111,11 +71,13 @@ const App = () => {
         closeActiveModal();
       }
     };
+
     const handleOverlayClick = (e) => {
       if (e.target.classList.contains("modal_opened")) {
         closeActiveModal();
       }
     };
+
     document.addEventListener("keydown", handleEscClose);
     document.addEventListener("mousedown", handleOverlayClick);
 
@@ -124,48 +86,25 @@ const App = () => {
       document.removeEventListener("mousedown", handleOverlayClick);
     };
   }, [activeModal]);
+
   return (
     <CurrentUserContext.Provider value={currentUser}>
       <CurrentLocationContext.Provider value={currentLocation}>
         <div className="page">
           <div className="page__section">
             <div className="page__content">
+              <Navigation
+                handleLogout={handleLogout}
+                // handleLoginClick={handleLoginClick} // you can enable if needed
+                isLoggedIn={isLoggedIn}
+              />
               <Routes>
                 <Route
                   path="/"
                   element={
-                    <>
-                      <Header
-                        handleLoginClick={handleSigninClick}
-                        isLoggedIn={isLoggedIn}
-                        onSearch={handleSearch}
-                        handleLogout={handleLogout}
-                      />
-                      {isLoading && <Preloader />}
-                      {!isLoading && error && <div>{error}</div>}
-                      {!isLoading && searchResults.length > 0 && (
-                        <NewsGrid
-                          searchResults={searchResults}
-                          onShowMore={handleShowMore}
-                          moreArticles={moreArticles}
-                        />
-                      )}
-                      {!isLoading &&
-                        searchResults.length === 0 &&
-                        searchQuery && (
-                          <div className="no-results">
-                            <img src={notfoundImg} alt="not found image" />
-                            <h1>Nothing found</h1>
-                            <p>
-                              Sorry but nothing matched your search item"
-                              {searchQuery}"
-                            </p>
-                          </div>
-                        )}
-                      <About />
-                    </>
+                    <Main isLoggedIn={isLoggedIn} handleLogout={handleLogout} />
                   }
-                ></Route>
+                />
                 <Route
                   path="/saved-news"
                   element={
@@ -175,7 +114,7 @@ const App = () => {
                       handleLogout={handleLogout}
                     />
                   }
-                ></Route>
+                />
               </Routes>
               <RegisterModal
                 isOpen={activeModal === "signup"}
