+++ src/components/Header/Header.jsx
@@ -3,14 +3,9 @@ import Navigation from "../Navigation/Navigation";
 import SearchForm from "../SearchForm/SearchForm";
 import "./Header.css";
 
-function Header({ handleLoginClick, isLoggedIn, onSearch, handleLogout }) {
+function Header({ onSearch }) {
   return (
     <div className="header">
-      <Navigation
-        handleLogout={handleLogout}
-        handleLoginClick={handleLoginClick}
-        isLoggedIn={isLoggedIn}
-      ></Navigation>
       <SearchForm onSearch={onSearch} />
     </div>
   );
